+++ src/chameleon/tests/test_templates.py
@@ -525,7 +525,18 @@ class ZopeTemplatesTestSuite(RenderTestCase):
             # source when debugging
             self.shortDescription = lambda: input_path
 
-            template = factory(input_path, keep_source=True, strict=False)
+            # Very implicitly enable implicit translation based on
+            # a string included in the input path:
+            implicit_i18n = 'implicit-i18n' in input_path
+            implicit_i18n_attrs = ("alt", "title") if implicit_i18n else ()
+
+            template = factory(
+                input_path,
+                keep_source=True,
+                strict=False,
+                implicit_i18n_translate=implicit_i18n,
+                implicit_i18n_attributes=implicit_i18n_attrs,
+                )
 
             params = kwargs.copy()
             params.update({